Fix precision

+++ crates/cubecl-linalg/src/matmul/components/spec.rs
@@ -47,9 +47,9 @@ impl MatmulPrecision for f16 {
 
 impl MatmulPrecision for f64 {
     const QUANTIZED: bool = false;
-    type EG = f16;
-    type ES = f16;
-    type EA = f32;
+    type EG = f64;
+    type ES = f64;
+    type EA = f64;
 }
 
 impl MatmulPrecision for flex32 {
@@ -61,9 +61,9 @@ impl MatmulPrecision for flex32 {
 
 impl MatmulPrecision for bf16 {
     const QUANTIZED: bool = false;
-    type EG = f64;
-    type ES = f64;
-    type EA = f64;
+    type EG = bf16;
+    type ES = bf16;
+    type EA = f32;
 }
 
 impl MatmulPrecision for f32 {
